requirements.txt
anthropic>=0.49.0
feedparser>=6.0.11
pyyaml>=6.0.0
